Reorganize execute code
