+++ test/execute/main.js
@@ -1,6 +1,6 @@
 import expect, {createSpy, spyOn} from 'expect'
 import xmock from 'xmock'
-import {execute, buildRequest, baseUrl, applySecurities, self as stubs} from '../src/execute'
+import {execute, buildRequest, baseUrl, applySecurities, self as stubs} from '../../src/execute'
 
 // Supported shape...  { spec, operationId, parameters, securities, fetch }
 // One can use operationId or pathItem + method
