+++ test/oas3/execute/main.js
@@ -4,7 +4,7 @@ import path from 'path'
 import fs from 'fs'
 import jsYaml from 'js-yaml'
 
-import {execute, buildRequest, baseUrl, applySecurities, self as stubs} from '../../src/execute'
+import {execute, buildRequest, baseUrl, applySecurities, self as stubs} from '../../../src/execute'
 
 const petstoreSpec = jsYaml.safeLoad(fs.readFileSync(path.join('test', 'oas3', 'data', 'petstore-oas3.yaml'), 'utf8'))
